Array keys/object properties should take precedence over global functs
Breaking change: Dash\property() no longer accepts callable paths.

Global functions were masking object properties with the same name. For example:

    Dash\get(['abs' => 'value'], 'abs')

was producing the wrong result because get() was calling the global abs() method instead of returning the value of the 'abs' property.

This was caused by the fact that Dash\property() short-circuited and returned callable paths right away, and then this returned value was called as a function instead of the intended property.

The fix was to remove the ability for Dash\property() to handle callables and push callable handling further up the call chain to map(), get(), etc.

Fixes #6