Dash::setCustom(): Remove optional auto-currying parameter
This allowed custom methods to be auto-curried by prefixing them with an underscore when invoking them on a chain. However, underscore-prefixing of curried functions was recently replaced with a new Dash\Curry namespace, so this behavior is no longer consistent with the built-in currying semantics.

To replicate this behavior on demand, use Dash\currify()

[breaking change]